refactor(security): deduplicate LoginFailureReason, remove csrf.middleware coupling
- Extract LoginFailureReason to common/security/login-failure-reason.ts;
  re-export from both login-audit.service and login-audit.helpers —
  eliminates the two divergent type definitions
- Inline pathWithoutQuery in throttler-exception.filter.ts so it no longer
  imports from csrf.middleware (removes unrelated cross-module dependency)
