+++ backend/src/common/security/login-failure-reason.ts
@@ -1,0 +1,7 @@
+export type LoginFailureReason =
+  | 'invalid_credentials'
+  | 'account_locked'
+  | 'captcha_failed'
+  | 'validation_error'
+  | 'rate_limited'
+  | 'internal_error';
